third_party/blink/web_tests/platform/generic/fast/ruby/ruby-line-height-expected.txt

Test that the CSS property 'line-height' is not applied to ruby base and annotation texts.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS [object HTMLParagraphElement] is non-null.
PASS getLineHeight('p') is "48px"
PASS [object HTMLElement] is non-null.
FAIL getLineHeight('r') should be normal. Was 48px.
PASS [object HTMLElement] is non-null.
PASS getLineHeight('t') is "normal"
PASS successfullyParsed is true

TEST COMPLETE


The line height of this is three times normal"line-height: 48px;", but the ruby should have 'line-height: normal'.
